fix: resolve clippy uninlined_format_args and other lints in multiple crates
Fixes in llama-chat-db, llama-chat-desktop-tools, llama-chat-tools, llama-chat-types:
- Inline format args across clipboard, audio, ocr, ui-automation, browser tools
- Remove unnecessary casts in windows_ui.rs
- Add missing Safety docs, allow type_complexity and upper_case_acronyms
- Fix div_ceil, int_plus_one, and other minor lints

+++ crates/llama-chat-desktop-tools/src/ui_automation_tools/windows_ui.rs
@@ -263,6 +263,7 @@ impl UiElementInfo {
 
 #[cfg(windows)]
 thread_local! {
+    #[allow(clippy::type_complexity)]
     static UI_ELEMENT_CACHE: std::cell::RefCell<std::collections::HashMap<(isize, String, String), (std::time::Instant, Vec<UiElementInfo>)>> = std::cell::RefCell::new(std::collections::HashMap::new());
 }
 
@@ -396,12 +397,12 @@ pub(crate) fn find_ui_elements_all(
             if let Ok(r) = unsafe { parent.CurrentBoundingRectangle() } {
                 if r.right > r.left && r.bottom > r.top {
                     results.push(UiElementInfo {
-                        cx: ((r.left + r.right) / 2) as i32,
-                        cy: ((r.top + r.bottom) / 2) as i32,
-                        left: r.left as i32,
-                        top: r.top as i32,
-                        width: (r.right - r.left) as i32,
-                        height: (r.bottom - r.top) as i32,
+                        cx: (r.left + r.right) / 2,
+                        cy: (r.top + r.bottom) / 2,
+                        left: r.left,
+                        top: r.top,
+                        width: r.right - r.left,
+                        height: r.bottom - r.top,
                         name: name.clone(),
                         control_type: control_type.clone(),
                     });
